added worker, intervaled GET requests
