+++ package.json
@@ -14,6 +14,7 @@
     "bull": "^4.8.3",
     "cesium": "^1.92.0",
     "dotenv": "^16.0.0",
-    "express": "^4.18.0"
+    "express": "^4.18.0",
+    "socket.io": "^4.5.1"
   }
 }
