+++ test-python.py
@@ -1,0 +1,1 @@
+print('hello world')
